+++ python/cugraph/dask/link_analysis/mg_pagerank_wrapper.pyx
@@ -47,15 +47,22 @@ def mg_pagerank(input_df,
 
     src = input_df['src']
     dst = input_df['dst']
+    vertex_t = src.dtype
+    if num_global_edges > (2**31 - 1):
+        edge_t = np.dtype("int64")
+    else:
+        edge_t = np.dtype("int32")
     if "value" in input_df.columns:
         weights = input_df['value']
-        weight_dtype = weights.dtype
+        weight_t = weights.dtype
     else:
-        weight_dtype = np.dtype("float32")
+        weight_t = np.dtype("float32")
 
     # FIXME: Offsets and indices are currently hardcoded to int, but this may
     #        not be acceptable in the future.
-    weightTypeMap = {np.dtype("float32") : <int>numberTypeEnum.floatType,
+    numberTypeMap = {np.dtype("int32") : <int>numberTypeEnum.int32Type,
+                     np.dtype("int64") : <int>numberTypeEnum.int64Type,
+                     np.dtype("float32") : <int>numberTypeEnum.floatType,
                      np.dtype("double") : <int>numberTypeEnum.doubleType}
 
     # FIXME: needs to be edge_t type not int
@@ -74,12 +81,13 @@ def mg_pagerank(input_df,
                              handle_[0],
                              <void*>c_src_vertices, <void*>c_dst_vertices, <void*>c_edge_weights,
                              <void*>c_vertex_partition_offsets,
-                             <numberTypeEnum>(<int>(numberTypeEnum.intType)),
-                             <numberTypeEnum>(<int>(numberTypeEnum.intType)),
-                             <numberTypeEnum>(<int>(weightTypeMap[weight_dtype])),
+                             <numberTypeEnum>(<int>(numberTypeMap[vertex_t])),
+                             <numberTypeEnum>(<int>(numberTypeMap[edge_t])),
+                             <numberTypeEnum>(<int>(numberTypeMap[weight_t])),
                              num_partition_edges,
                              num_global_verts, num_global_edges,
                              partition_row_size, partition_col_size,
+                             True,
                              True, True) 
 
     df = cudf.DataFrame()